pyproject.toml
[project]
name = "hack3school"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "prisma>=0.15.0",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
    "upyloadthing>=0.2.1",
]
